Smart check for render method
If we do it like this, we save so much code. This simply lets us call the render method on the dataTable from the `SmartDataTables` middleware without any extra check.

So simply, if none of the checks return true and we dont get a response by line 163 and if the dataTable is not smart, we can return false.

We will use this false in the middleware like:
```php

        $dataTable = app($request->get('smartDataTable'));

        if ($response = $dataTable->render()) {
        	return $response;
        }

        return $next($request);
```